Add files via upload
A simple project which displays the current date in Northern Sámi, Old Norse and Swedish. 

This project was done in order to practice my JavasScript skills as well as my SáN and ON skills. I have taken some creative liberties with the ON translation as Norse society lacked a uniform month-based calendar.  Enjoy!